- Refactor, merge UserInput into UserInterface class - Remove dependency of Service to UserInterface class - Add category table
