--- Phonebook.SpyrosZoupas/Phonebook.SpyrosZoupas/DAL/Contact.cs
@@ -1,10 +1,0 @@
-﻿namespace Phonebook.SpyrosZoupas.DAL
-{
-    public class Contact
-    {
-        public int Id { get; set; }
-        public string Name { get; set; }
-        public string Email { get; set; }
-        public string PhoneNumber { get; set; }
-    }
-}
