+++ Phonebook.SpyrosZoupas/Phonebook.SpyrosZoupas/DAL/ContactContext.cs
@@ -1,4 +1,5 @@
 ﻿using Microsoft.EntityFrameworkCore;
+using Phonebook.SpyrosZoupas.DAL.Models;
 
 namespace Phonebook.SpyrosZoupas.DAL
 {
@@ -6,6 +7,8 @@ namespace Phonebook.SpyrosZoupas.DAL
     public class ContactContext : DbContext // represents a session with the DB allows you to query the DB
     {
         public DbSet<Contact> Contacts { get; set; } // represents a Contact table. Contact model is a row of Contact table
+
+        public DbSet<Category> Categories { get; set; }
 
         protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder) =>
             optionsBuilder.UseSqlServer($"Server=(LocalDb)\\TheCSharpAcademy;Database=Phonebook;Trusted_Connection=True;");
