+++ Phonebook.SpyrosZoupas/Phonebook.SpyrosZoupas/DAL/Models/Category.cs
@@ -1,0 +1,15 @@
+﻿using Microsoft.EntityFrameworkCore;
+using System.ComponentModel.DataAnnotations;
+
+namespace Phonebook.SpyrosZoupas.DAL.Models
+{
+    [Index(nameof(Name), IsUnique = true)]
+    public class Category
+    {
+        [Key]
+        public int CategoryId { get; set; }
+        [Required]
+        public string Name { get; set; }
+        public List<Contact> Contacts { get; set; }
+    }
+}
