+++ Phonebook.SpyrosZoupas/Phonebook.SpyrosZoupas/Enums.cs
@@ -1,0 +1,16 @@
+﻿namespace Phonebook.SpyrosZoupas
+{
+    public class Enums
+    {
+        public enum MenuOptions
+        {
+            AddContact,
+            DeleteContact,
+            UpdateContact,
+            ViewContact,
+            ViewAllContacts,
+            SendEmail,
+            Quit
+        }
+    }
+}
