+++ Phonebook.SpyrosZoupas/Phonebook.SpyrosZoupas/UserInterface.cs
@@ -1,13 +1,65 @@
-﻿using Phonebook.SpyrosZoupas.DAL;
+﻿using Phonebook.SpyrosZoupas.DAL.Models;
 using Spectre.Console;
+using static Phonebook.SpyrosZoupas.Enums;
 
 namespace Phonebook.SpyrosZoupas
 {
     public class UserInterface
     {
+        private readonly ContactService _contactService;
+
+        public UserInterface(ContactService contactService)
+        {
+            _contactService = contactService;
+        }
+
+        public void MainMenu()
+        {
+            var isAppRunning = true;
+            while (isAppRunning)
+            {
+                var option = AnsiConsole.Prompt(
+                new SelectionPrompt<MenuOptions>()
+                .Title("What would you like to do?")
+                .AddChoices(
+                    MenuOptions.AddContact,
+                    MenuOptions.DeleteContact,
+                    MenuOptions.UpdateContact,
+                    MenuOptions.ViewAllContacts,
+                    MenuOptions.ViewContact,
+                    MenuOptions.SendEmail,
+                    MenuOptions.Quit));
+
+                switch (option)
+                {
+                    case MenuOptions.AddContact:
+                        _contactService.InsertContact();
+                        break;
+                    case MenuOptions.DeleteContact:
+                        _contactService.DeleteContact();
+                        break;
+                    case MenuOptions.UpdateContact:
+                        _contactService.UpdateContact();
+                        break;
+                    case MenuOptions.ViewContact:
+                        ShowContact(_contactService.GetContact());
+                        break;
+                    case MenuOptions.ViewAllContacts:
+                        ShowContactTable(_contactService.GetAllContacts());
+                        break;
+                    case MenuOptions.SendEmail:
+                        _contactService.SendEmail();
+                        break;
+                    case MenuOptions.Quit:
+                        Environment.Exit(0);
+                        break;
+                }
+            }
+        }
+
         public void ShowContact(Contact contact)
         {
-            var panel = new Panel($@"Id: {contact.Id}
+            var panel = new Panel($@"Id: {contact.ContactId}
 Name: {contact.Name}
 Email: {contact.Email}
 Phone Number: {contact.PhoneNumber}");
@@ -31,7 +83,7 @@ Phone Number: {contact.PhoneNumber}");
 
             foreach (Contact contact in contacts)
             {
-                table.AddRow(contact.Id.ToString(), contact.Name, contact.Email, contact.PhoneNumber);
+                table.AddRow(contact.ContactId.ToString(), contact.Name, contact.Email, contact.PhoneNumber);
             }
 
             AnsiConsole.Write(table);
